Movement sets to zero on collision
The knight (knight sprite represents the main character) cannot move further when hitting the 'Base' tagged collider. Using the OnCollisionEnter2D function which  is Unity callback method that is part of the MonoBehaviour class and is used to detect and respond to collisions in a 2D physics environment. Check if the collider object has 'Base" tag. If it does, set the velocity of the rigidbody to zero and this will prevent the character from moving.