Enemy habitat (range)
If the player is in the range of the enemy, enemy follows the player. If the player goes out of the range, enemy returns back to its habitat position i.e homePos.

Also changed the grid size to 1 from 0.32, which made easier to calculate the distance between two objects and their transform position. And also all the sprite size from 100 pixels per unit to 32 pixels per unit.