Player gains experience and level up increasing some stats
ExperienceManager is a singleton class and can be call from anywhere in the game, that has AddExperience(int amount) method which invokes OnExperienceChange event.

Added an expAmount integer in the enemyHealthManager script. When the enemy object is disabled (player defeats this enemy), the Enemny.Instance.AddExperience(expAmount) will invoke the OnExperienceChange event and add the expAmount.

The character script is attached to the player object ( in my project Warrior ). Everytime the event gets called, the HandleExperienceChange(int newExperience) gets called with the amount.

This way player can level up.

The other changes were just arranging  the scripts and sprites in proper folders.