Some fixes and Made a Fog effect using shader graph
Updates and fixes:

NPC1 script for the first NPC to be excluded from the scene after interaction.

In DialogManager script starts a coroutine after the conversation ends and screen fades to black. After fade out the NPC object is disabled.

Loading and saving of the NPC state:

npcObject.SetActive(!data.isNPCDisabled):
data.isNPCDisabled represents whether the NPC game object should be disabled (true) or not (false) based on the saved game data.
The logical NOT operator ! is used to invert the value of data.isNPCDisabled. If isNPCDisabled is true, it means the NPC should be disabled, so !true evaluates to false, and vice versa.
SetActive() method is then called on npcObject with the inverted value. If isNPCDisabled was true, SetActive(false) is called, disabling the NPC. If it was false, SetActive(true) is called, enabling the NPC.

data.isNPCDisabled = !npcObject.activeSelf:
npcObject.activeSelf returns true if the NPC game object is currently active (not disabled) in the scene, and false if it's inactive (disabled).
The logical NOT operator ! is used to invert the value of npcObject.activeSelf. If the NPC is currently active, !true evaluates to false, indicating that the NPC should not be disabled (isNPCDisabled should be false). If the NPC is inactive, !false evaluates to true, indicating that the NPC should be disabled (isNPCDisabled should be true).